refactor(orchestrator): move scheduling logic to TimeoutProcessBatchOperation and remove wrapper service
